Add unit and bridge code
Add some code to deal with different possibilites of fundamental
unit.

This does not currently deal with the fact that magma's
FundamentalUnit(K) returns inconsistent signs, which changes our
results.

Also:

* wrote a function that checks all the split primes for roots
mod p, the main point of our code
* Added "authors" comment
* added pseudocode for a main function
* fixed some whitespace issues
* factored out a few computations from S() and C()
into the CheckField() function, so that we don't have
to do those each time in the loop.