Enforce positive trace for fundamental unit
The FundamentalUnit() function from Magma
is inconsistent on the signs it gives the
real and square root part of the unit.

I haven't found an easy way to access these
coefficients, which is maddening.
But, for our computations, the "rational" part
is the only one that matters, so in this
commit I added functionality to get around
this by enforcing that the trace must be positive.